Add development env guards

+++ app/models/concerns/purgeable.rb
@@ -7,10 +7,14 @@ module Purgeable
 
   module ClassMethods
     def purge_all
+      return if Rails.env.development?
+
       service.purge_by_key(table_key)
     end
 
     def soft_purge_all
+      return if Rails.env.development?
+
       service.purge_by_key(table_key, true)
     end
 
@@ -19,10 +23,14 @@ module Purgeable
     end
 
     def fastly
+      return if Rails.env.development?
+
       Fastly.new(api_key: ApplicationConfig["FASTLY_API_KEY"])
     end
 
     def service
+      return if Rails.env.development?
+
       Fastly::Service.new({ id: ApplicationConfig["FASTLY_SERVICE_ID"] }, fastly)
     end
   end
@@ -37,10 +45,14 @@ module Purgeable
   end
 
   def purge
+    return if Rails.env.development?
+
     service.purge_by_key(record_key)
   end
 
   def soft_purge
+    return if Rails.env.development?
+
     service.purge_by_key(record_key, true)
   end
